HPC Pack 2019 integration
1. Get rid of DSC resources, directly use script
2. Upgrade to HPC Pack 2019
3. Fix some issues in autoscaler
