+++ project.ini
@@ -3,7 +3,7 @@ version = 1.2.0
 name = hpcpack
 
 [blobs hpcpack]
-Files = nuget.exe, hpc-comm.pfx, hpc-comm.pem, vnet.json, ndp48-web.exe, HPCPack2016Update3-Full-Refresh-v6450.zip, cyclecloud-hpcpack-pkg-1.2.0.zip
+Files = nuget.exe, hpc-comm.pfx, hpc-comm.pem, vnet.json, ndp48-web.exe, HPCComputeNode-7214.zip, cyclecloud-hpcpack-pkg-1.2.0.zip
 # Files = nuget.exe, hpc-comm.pfx, hpc-comm.pem, vnet.json, ndp48-web.exe, HPCPack2016Update3-Full-Refresh-v6450.zip, cyclecloud_api-8.0.1-py2.py3-none-any.whl, cyclecloud-scalelib-0.1.1.tar.gz, hpcpack-autoscaler.zip
 
 [spec hn]
