--- specs/default/chef/site-cookbooks/hpcpack/files/dsc-extension.json
@@ -1,51 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "vmName": {
-      "type": "string",
-      "metadata": {
-        "description": "The VM name"
-      }
-    },
-    "dscExtensionName": {
-      "type": "string",
-      "defaultValue": "configNodeWithDsc",
-      "metadata": {
-        "description": "The name of the Dsc extension"
-      }
-    },
-    "dscSettings": {
-      "type": "object",
-      "metadata": {
-        "description": "The DSC public settings"
-      }
-    },
-    "dscProtectedSettings": {
-      "type": "secureObject",
-      "defaultValue": {
-      },
-      "metadata": {
-        "description": "The DSC protected settings"
-      }
-    }
-  },
-  "variables": {
-  },
-  "resources": [
-    {
-      "apiVersion": "2019-03-01",
-      "type": "Microsoft.Compute/virtualMachines/extensions",
-      "name": "[concat(parameters('vmName'), '/', parameters('dscExtensionName'))]",
-      "location": "[resourceGroup().location]",
-      "properties": {
-        "publisher": "Microsoft.Powershell",
-        "type": "DSC",
-        "typeHandlerVersion": "2.80",
-        "autoUpgradeMinorVersion": true,
-        "settings": "[parameters('dscSettings')]",
-        "protectedSettings": "[parameters('dscProtectedSettings')]"
-      }
-    }
-  ]
-}
